Multiattachments shareFromOutside

+++ app/src/main/java/com/naposystems/napoleonchat/source/remote/dto/validateMessageEvent/ValidateMessageEventDTO.kt
@@ -12,6 +12,6 @@ data class ValidateMessageEventDTO(
 @JsonClass(generateAdapter = true)
 data class ValidateMessage(
     @Json(name = "id") val id: String,
-    @Json(name = "user") val user: Int,
+    @Json(name = "user") val user: Long,
     @Json(name = "status") val status: Int
 )
